fix: filenames
- resolves `TS1261` errors
- aligns filenames with standards and import statements

- `account.svg` -> `Account.svg`
- `accountTypes.ts` -> `AccountTypes.ts`
- `admin.png` -> `Admin.png`
- `arrowsconnect.svg` -> `ArrowsConnect.svg`
- `benefit.svg -> Benefit.svg`
- `card1.png` -> `Card1.png`
- `chatframe.svg` -> `ChatFrame.svg`
- `chevronleftdouble.svg` -> `ChevronLeftDouble.svg`
- `chevronrightdouble.svg` -> `ChevronRightDouble.svg`
- `discordIcon.png` -> `DiscordIcon.png`
- `District53intro.tsx` -> `District53Intro.tsx`
- `districteffect.svg` -> `DistrictEffect.svg`
- `districteffect1.svg` -> `DistrictEffect1.svg`
- `districteffect2.svg` -> `DistrictEffect2.svg`
- `donotdisturbframe.svg` -> `DoNotDisturbFrame.svg`
- `downloadbig.svg` -> `DownloadBig.svg`
- `downloadsmall.svg` -> `DownloadSmall.svg`
- `ellipse.svg` -> `Ellipse.svg`
- `emotion.svg` -> `Emotion.svg`
- `home.svg` -> `Home.svg`
- `homeTypes.ts` -> `HomeTypes.ts`
- `library.svg` -> `Library.svg`
- `Librarymode.tsx` -> `LibraryMode.tsx`
- `Libraryshow.tsx` -> `LibraryShow.tsx`
- `linkIcon.png` -> `LinkIcon.png`
- `linux.svg` -> `Linux.svg`
- `mac.svg` -> `Mac.svg`
- `mail.png -> Mail.png`
- `mask.png -> Mask.png`
- `maximize.svg` -> `Maximize.svg`
- `minimize.svg` -> `Minimize.svg`
- `moderator.png` -> `Moderator.png`
- `newlogo.png` -> `NewLogo.png`
- `newlogohead.png` -> `NewLogoHead.png`
- `nocontact.png` -> `NoContact.png`
- `nogames.png` -> `NoGames.png`
- `noreviews.png` -> `NoReviews.png`
- `offlineframe.svg` -> `OfflineFrame.svg`
- `onlineframe.svg` -> `OnlineFrame.svg`
- `property.png` -> `Property.png`
- `question.svg -> Question.svg`
- `searchlg.svg` -> `SearchLg.svg`
- `settings.svg` -> `Settings.svg`
- `settingTypes.ts` -> `SettingTypes.ts`
- `solarblockchain.png` -> `SolarBlockchain.png`
- `store.svg` -> `Store.svg`
- `walletTypes.ts` -> `WalletTypes.ts`
- `warning.svg -> Warning.svg`
- `windows.png` -> `Windows.png`
- `youtubeIcon.png` -> `YoutubeIcon.png`
