+++ src/assets/main/SolarBlockchain.png
